+++ fbw-common/src/wasm/systems/systems/src/hydraulic/mod.rs
@@ -43,6 +43,7 @@ pub mod linear_actuator;
 pub mod nose_steering;
 pub mod pumps;
 pub mod pushback;
+pub mod reverser;
 pub mod rudder_control;
 pub mod trimmable_horizontal_stabilizer;
 
@@ -3123,6 +3124,100 @@ impl PressureSource for ManualPump {
 }
 impl HeatingElement for ManualPump {}
 impl HeatingPressureSource for ManualPump {}
+
+#[derive(PartialEq, Clone, Copy)]
+enum HydraulicValveType {
+    ClosedWhenOff,
+    _OpenedWhenOff,
+    Mechanical,
+}
+
+struct HydraulicValve {
+    position: LowPassFilter<Ratio>,
+    is_powered: bool,
+    powered_by: Option<Vec<ElectricalBusType>>,
+    valve_type: HydraulicValveType,
+
+    pressure_input: Pressure,
+    pressure_output: Pressure,
+}
+impl HydraulicValve {
+    const POSITION_RESPONSE_TIME_CONSTANT: Duration = Duration::from_millis(150);
+    const MIN_POSITION_FOR_ZERO_PRESSURE_RATIO: f64 = 0.02;
+
+    fn new(valve_type: HydraulicValveType, powered_by: Option<Vec<ElectricalBusType>>) -> Self {
+        Self {
+            position: LowPassFilter::<Ratio>::new(Self::POSITION_RESPONSE_TIME_CONSTANT),
+            is_powered: false,
+            powered_by,
+            valve_type,
+            pressure_input: Pressure::default(),
+            pressure_output: Pressure::default(),
+        }
+    }
+
+    fn update(
+        &mut self,
+        context: &UpdateContext,
+        commanded_open: bool,
+        current_pressure_input: Pressure,
+    ) {
+        let commanded_position = self.actual_target_position_from_valve_type(commanded_open);
+
+        self.position.update(context.delta(), commanded_position);
+
+        self.pressure_input = current_pressure_input;
+        self.update_output_pressure();
+    }
+
+    fn actual_target_position_from_valve_type(&self, commanded_open: bool) -> Ratio {
+        match self.valve_type {
+            HydraulicValveType::_OpenedWhenOff => {
+                if !commanded_open && self.is_powered {
+                    Ratio::new::<ratio>(0.)
+                } else {
+                    Ratio::new::<ratio>(1.)
+                }
+            }
+            HydraulicValveType::ClosedWhenOff => {
+                if commanded_open && self.is_powered {
+                    Ratio::new::<ratio>(1.)
+                } else {
+                    Ratio::new::<ratio>(0.)
+                }
+            }
+            HydraulicValveType::Mechanical => {
+                if commanded_open {
+                    Ratio::new::<ratio>(1.)
+                } else {
+                    Ratio::new::<ratio>(0.)
+                }
+            }
+        }
+    }
+
+    fn update_output_pressure(&mut self) {
+        self.pressure_output =
+            if self.position.output().get::<ratio>() > Self::MIN_POSITION_FOR_ZERO_PRESSURE_RATIO {
+                self.pressure_input
+                    * (self.position.output().sqrt() * 1.4)
+                        .min(Ratio::new::<ratio>(1.).max(Ratio::new::<ratio>(0.)))
+            } else {
+                Pressure::default()
+            }
+    }
+
+    fn pressure_output(&self) -> Pressure {
+        self.pressure_output
+    }
+}
+impl SimulationElement for HydraulicValve {
+    fn receive_power(&mut self, buses: &impl ElectricalBuses) {
+        if let Some(powered_by_element) = &self.powered_by {
+            self.is_powered = buses.any_is_powered(powered_by_element);
+        }
+    }
+}
 
 #[cfg(test)]
 mod tests {