+++ .github/CHANGELOG.md
@@ -14,6 +14,7 @@
 1. [SD] Corrected fuel transfer valve amber logic - @tracernz (Mike)
 1. [FMS] TO SPEEDS TOO LOW uses FQI fuel quantity after engine start to avoid spurious warnings - @tracernz (Mike)
 1. [ND] Fix ROSE VOR/LS conditions and ILS course - @tracernz (Mike)
+1. [HYD] Added hydraulic reverser actuators - @Crocket63 (crocket)
 
 ## 0.10.0
 
@@ -139,6 +140,7 @@
 1. [EFB] Added pause at T/D function - @2hwk (2Cas#1022)
 1. [LIGHTS] Fixed trim decal emissive and floods - @FinalLightNL (FinalLight#2113)
 1. [FLIGHTMODEL/FUEL] Fix outer tank transfer behaviour - @donstim (donbikes#4084)
+
 
 ## 0.9.0